refactor: refactor component `Popover` and `Tooltip`
