+++ src/components/popover.js
@@ -1,20 +1,13 @@
-import { createPopper } from '@popperjs/core';
+import BasePopper from './basePopper';
 
 const Default = {
-    placement: 'top',
     offset: 10,
-    triggerType: 'hover',
-    onShow: () => { },
-    onHide: () => { }
 }
 
-class Popover {
+class Popover extends BasePopper {
     constructor(targetEl = null, triggerEl = null, options = {}) {
-        this._targetEl = targetEl
-        this._triggerEl = triggerEl
-        this._options = { ...Default, ...options }
-        this._popperInstance = this._createPopperInstace()
-        this._init()
+        super(targetEl, triggerEl, options)
+        this._options = { ...Default, ...this._options }
     }
 
     _init() {
@@ -45,77 +38,6 @@ class Popover {
                 })
             })
         }
-    }
-
-    _createPopperInstace() {
-        return createPopper(this._triggerEl, this._targetEl, {
-            placement: this._options.placement,
-            modifiers: [
-                {
-                    name: 'offset',
-                    options: {
-                        offset: [0, this._options.offset],
-                    },
-                },
-            ],
-        });
-    }
-
-    _getTriggerEvents() {
-        switch (this._options.triggerType) {
-            case 'hover':
-                return {
-                    showEvents: ['mouseenter', 'focus'],
-                    hideEvents: ['mouseleave', 'blur']
-                }
-            case 'click':
-                return {
-                    showEvents: ['click', 'focus'],
-                    hideEvents: ['focusout', 'blur']
-                }
-            default:
-                return {
-                    showEvents: ['mouseenter', 'focus'],
-                    hideEvents: ['mouseleave', 'blur']
-                }
-        }
-    }
-
-    show() {
-        this._targetEl.classList.remove('opacity-0', 'invisible')
-        this._targetEl.classList.add('opacity-100', 'visible')
-
-        // Enable the event listeners
-        this._popperInstance.setOptions(options => ({
-            ...options,
-            modifiers: [
-                ...options.modifiers,
-                { name: 'eventListeners', enabled: true },
-            ],
-        }));
-
-        // Update its position
-        this._popperInstance.update()
-
-        // callback function
-        this._options.onShow(this)
-    }
-
-    hide() {
-        this._targetEl.classList.remove('opacity-100', 'visible')
-        this._targetEl.classList.add('opacity-0', 'invisible')
-
-        // Disable the event listeners
-        this._popperInstance.setOptions(options => ({
-            ...options,
-            modifiers: [
-                ...options.modifiers,
-                { name: 'eventListeners', enabled: false },
-            ],
-        }));
-
-        // callback function
-        this._options.onHide(this)
     }
 }
